+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- markdown output to playground
+
 ### Changed
 
 - base playground docker image set to eclipse-temurin:21_35-jre